rename kbd utils
