+++ packages/tldraw/src/lib/ui/components/Toolbar/DefaultToolbar.tsx
@@ -9,7 +9,7 @@ import { useReadonly } from '../../hooks/useReadonly'
 import { TLUiToolbarItem, useToolbarSchema } from '../../hooks/useToolbarSchema'
 import { TLUiToolItem } from '../../hooks/useTools'
 import { useTranslation } from '../../hooks/useTranslation/useTranslation'
-import { kbdStr } from '../../shared'
+import { kbdStr } from '../../kbd-utils'
 import { MobileStylePanel } from '../MobileStylePanel'
 import { TldrawUiButton } from '../primitives/Button/TldrawUiButton'
 import { TldrawUiButtonIcon } from '../primitives/Button/TldrawUiButtonIcon'
