+++ packages/tldraw/src/lib/ui/kbd-utils.ts
@@ -1,11 +1,3 @@
-/** @internal */
-export function toStartCase(str: string) {
-	return str
-		.split(' ')
-		.map((s) => s.charAt(0).toUpperCase() + s.slice(1))
-		.join(' ')
-}
-
 const isDarwin =
 	typeof window === 'undefined'
 		? false
@@ -13,7 +5,7 @@ const isDarwin =
 const cmdKey = isDarwin ? '⌘' : 'Ctrl'
 const altKey = isDarwin ? '⌥' : 'Alt'
 
-/** @internal */
+/** @public */
 export function kbd(str: string) {
 	return str
 		.split(',')[0]
@@ -24,7 +16,7 @@ export function kbd(str: string) {
 		})
 }
 
-/** @internal */
+/** @public */
 export function kbdStr(str: string) {
 	return '— ' + kbd(str).join(' ')
 }
